requirements.txt
asgiref==3.4.1
Django==3.2.23
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.4.0
importlib-metadata==4.8.3
Markdown==3.3.7
mysqlclient==2.1.1
Pillow==8.4.0
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.1.1
tzdata==2023.4
zipp==3.6.0
